Hopefully this will catch everything. I may still be missing a few files so wait a bit before updating.
LOTS of changes. Elicit can now save/load swatches and shots. Just click on the main swatch/shot to add it to the respective panels. click on the ones in the panel to load them back up. There is still some work to be done, but i'm not sure when i'll be able to get to it, so i figured i'd commit what i have :)

Currently only the winter theme has been updated to use the new features.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk/e17/apps/elicit@10175 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33